requirements.txt
fastapi[all]~=0.112.2
pyjwt~=2.9.0
passlib[bcrypt]~=1.7.4
sqlalchemy~=2.0.32
python-dotenv~=1.0.1
pydantic~=2.8.2